BS-167 | Adding Privilege to access the service
