+++ api/src/main/resources/moduleApplicationContext.xml
@@ -11,4 +11,30 @@
   		    http://www.springframework.org/schema/context/spring-context-3.0.xsd">
     
     <context:component-scan base-package="org.openmrs.module.fhirExtension"/>
+
+    <bean id="exportTask" class="org.springframework.transaction.interceptor.TransactionProxyFactoryBean">
+        <property name="transactionManager">
+            <ref bean="transactionManager"/>
+        </property>
+        <property name="preInterceptors">
+            <ref bean="serviceInterceptors"/>
+        </property>
+        <property name="transactionAttributeSource">
+            <ref bean="transactionAttributeSource"/>
+        </property>
+        <property name="target">
+            <bean class="org.openmrs.module.fhirExtension.service.impl.ExportTaskImpl">
+                <constructor-arg ref="fhirTaskDaoImpl"/>
+            </bean>
+        </property>
+    </bean>
+
+    <bean parent="serviceContext">
+        <property name="moduleService">
+            <list merge="true">
+                <value>org.openmrs.module.fhirExtension.service.ExportTask</value>
+                <ref bean="exportTask"/>
+            </list>
+        </property>
+    </bean>
 </beans>
