+++ api/src/test/java/org/openmrs/module/fhirExtension/service/impl/ExportTaskTest.java
@@ -1,4 +1,4 @@
-package org.openmrs.module.fhirExtension.service;
+package org.openmrs.module.fhirExtension.service.impl;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -12,21 +12,15 @@ import org.openmrs.api.context.Context;
 import org.openmrs.module.fhir2.api.FhirConditionService;
 import org.openmrs.module.fhir2.api.dao.FhirTaskDao;
 import org.openmrs.module.fhir2.model.FhirTask;
-import org.openmrs.module.fhirExtension.export.Exporter;
-import org.openmrs.module.fhirExtension.export.impl.ConditionExport;
+import org.openmrs.module.fhirExtension.service.ExportToFileService;
 import org.powermock.api.mockito.PowerMockito;
 import org.powermock.core.classloader.annotations.PowerMockIgnore;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.mockito.Matchers.any;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @RunWith(PowerMockRunner.class)
@@ -37,17 +31,8 @@ public class ExportTaskTest {
 	@Mock
 	private FhirTaskDao fhirTaskDao;
 	
-	@Mock
-	private ConceptService conceptService;
-	
-	@Mock
-	private ExportToFileService exportToFileService;
-	
-	@Mock
-	private FhirConditionService fhirConditionService;
-	
 	@InjectMocks
-	private ExportTask exportTask;
+	private ExportTaskImpl exportTask;
 	
 	@Before
 	public void setUp() {
@@ -65,31 +50,6 @@ public class ExportTaskTest {
 		
 		assertNotNull(initialTaskResponse);
 		assertEquals(FhirTask.TaskStatus.ACCEPTED, initialTaskResponse.getStatus());
-	}
-	
-	@Test
-	public void shouldExportPatientDataAndUpdateFhirTaskStatusToCompleted_whenValidDateRangeProvided() {
-		FhirTask fhirTask = mockFhirTask();
-		
-		exportTask.export(fhirTask, "2023-01-01", "2023-12-31", Context.getUserContext(), "");
-		
-		assertEquals(FhirTask.TaskStatus.COMPLETED, fhirTask.getStatus());
-		verify(conceptService, times(1)).getConceptByName("Download URL");
-		verify(fhirTaskDao, times(1)).createOrUpdate(any(FhirTask.class));
-	}
-	
-	@Test (expected = RuntimeException.class)
-	public void shouldThrowException_whenInvalidDateRangeProvided() {
-		List<Exporter> exporters = new ArrayList<>();
-		exporters.add(new ConditionExport(fhirConditionService));
-		when(Context.getRegisteredComponents(Exporter.class)).thenReturn(exporters);
-
-		FhirTask fhirTask = mockFhirTask();
-
-		exportTask.export(fhirTask, "2023-AB-CD", "2023-12-31", Context.getUserContext(), "");
-
-		assertEquals(FhirTask.TaskStatus.REJECTED, fhirTask.getStatus());
-		verify(fhirTaskDao, times(1)).createOrUpdate(any(FhirTask.class));
 	}
 	
 	private FhirTask mockFhirTask() {
